+++ src/include/kernel.hpp
@@ -11,7 +11,7 @@
 #include <tuple/tuple.hpp>
 
 // data types and layout
-using type = float;
+using type = double;
 /*
 using type_x = type;
 using type_y = type;
@@ -30,6 +30,7 @@ using type_y = std::uint32_t;
 using type_z = std::uint32_t;
 using element_type = fw::tuple<type_x, type_y, type_z>;
 */
+
 using type_x = std::uint16_t;
 using type_y = double;
 using type_z = std::uint32_t;
